Use Int64 for carrier lengths and dimensions (#79)
Use Int64 for CArray dimensions and CString lengths. Recognize signed
32- and 64-bit carrier integer fields structurally, emit matching Python
ctypes, and guard 32-bit conversions against overflow. Use Int for
imported primitive layout metadata.

+++ JuliaLibWrapping/test/expected_carray3_lowlevel.py
@@ -56,7 +56,7 @@ _jlw_loaded.add(_jlw_this_pkg)
 
 class CArray_borrowed_Float64_3(ctypes.Structure):
     _fields_ = [
-        ("dims", (ctypes.c_int32 * 3)),
+        ("dims", (ctypes.c_int64 * 3)),
         ("data", ctypes.POINTER(ctypes.c_double)),
     ]
 
@@ -78,7 +78,7 @@ class CArray_borrowed_Float64_3(ctypes.Structure):
         expected_dtype = np.dtype("float64")
         if arr.dtype != expected_dtype:
             raise ValueError(f"expected dtype float64, got {arr.dtype}")
-        obj = cls(dims=(ctypes.c_int32 * 3)(*arr.shape),
+        obj = cls(dims=(ctypes.c_int64 * 3)(*arr.shape),
                   data=arr.ctypes.data_as(ctypes.POINTER(ctypes.c_double)))
         obj._buffer = arr
         return obj
